long session. specs + impl. components. Read more.
Enough for this project. If I have some time I will make a text command
parser and a history, but that is far from the specific knowledge of
react. I've left comment blocks around the project. I don't usually do
it, but I consider it appropriate to explain the decisions made in this
test.